+++ database/migrations/2025_03_29_174728_create_write_off_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('product_id')->constrained('products')->cascadeOnUpdate()->cascadeOnDelete();
             $table->foreignId('write_off_id')->constrained('write_offs')->cascadeOnUpdate()->cascadeOnDelete();
-            $table->integer('count');
+            $table->decimal('count');
             $table->foreignId('establishment_id')->constrained('establishments')->cascadeOnUpdate()->cascadeOnDelete();
             $table->timestamps();
         });
